+++ CHANGELOG.md
@@ -12,6 +12,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   canonical statement (see esutil.getCanonicalStatement), it will lookup
   the feature at that statement. This will be used by internal
   APIs to do scope-based dereferencing of super classes, behaviors, and mixins.
+* Added JavascriptImport#specifier for getting the original import specifier,
+  before it may have been resolved to a file-relative url by the node module
+  resolution algorithm.
 <!-- Add new, unreleased changes here. -->
 
 ## [3.0.0-pre.14] - 2018-03-09